+++ wp-content/plugins/ssl-insecure-content-fixer/includes/class.SSLInsecureContentFixer.php
@@ -37,9 +37,9 @@ class SSLInsecureContentFixer {
 		$this->proxyFix();
 		$this->configureSiteOnly();
 
-		add_action('init', array($this, 'init'));
+		add_action('init', array($this, 'loadTranslations'));
 
-		if ($this->options['fix_level'] !== 'off' && is_ssl()) {
+		if ($this->allowFixer()) {
 			add_action('init', array($this, 'runFilters'), 4);
 
 			// filter script and stylesheet links
@@ -109,6 +109,43 @@ class SSLInsecureContentFixer {
 			require SSLFIX_PLUGIN_ROOT . 'includes/class.SSLInsecureContentFixerAdmin.php';
 			new SSLInsecureContentFixerAdmin();
 		}
+	}
+
+	/**
+	* see whether the fixer should be run for this request
+	* @return bool
+	*/
+	protected function allowFixer() {
+		// do nothing if fixer is turned off
+		if ($this->options['fix_level'] === 'off') {
+			return false;
+		}
+
+		// don't mess with WooCommerce downloads
+		if (isset($_GET['download_file']) && isset($_GET['order']) && (isset($_GET['email']) || isset($_GET['uid']))) {
+			// but ensure that WooCommerce is active and will handle this request
+			if ($this->isPluginActive('woocommerce/woocommerce.php')) {
+				return false;
+			}
+		}
+
+		return is_ssl();
+	}
+
+	/**
+	* test whether a plugin is active
+	* @param string $plugin
+	* @return bool
+	*/
+	protected function isPluginActive($plugin) {
+		if (is_multisite()) {
+			$plugins = (array) get_site_option('active_sitewide_plugins', array());
+			if (isset($plugins[$plugin])) {
+				return true;
+			}
+		}
+
+		return in_array($plugin, (array) get_option('active_plugins', array()));
 	}
 
 	/**
@@ -266,7 +303,7 @@ class SSLInsecureContentFixer {
 	/**
 	* load text translations
 	*/
-	public function init() {
+	public function loadTranslations() {
 		load_plugin_textdomain('ssl-insecure-content-fixer');
 	}
 
@@ -291,6 +328,9 @@ class SSLInsecureContentFixer {
 			'#<img [^>]+srcset=["\']\K[^"\']+#is',					// responsive image srcset links (to external images; WordPress already handles local images)
 		);
 		$content = preg_replace_callback($embed_searches, array($this, 'fixContent_embed_callback'), $content);
+
+		// fix responsive images with data-* attributes
+		$content = preg_replace_callback('#<img [^>]*data-[^\'"]+=["\'][^>]+#is', array($this, 'fixContent_data_attr_callback'), $content);
 
 		return $content;
 	}
@@ -326,6 +366,17 @@ class SSLInsecureContentFixer {
 	public function fixContent_embed_callback($matches) {
 		// match from start of http: URL until either end quotes, space, or query parameter separator, thus allowing for URLs in parameters
 		$content = preg_replace_callback('#http://[^\'"&\? ]+#i', array($this, 'fixContent_src_callback'), $matches[0]);
+
+		return $content;
+	}
+
+	/**
+	* callback for fixContent() regex replace for data attributes on images
+	* @param array $matches
+	* @return string
+	*/
+	public function fixContent_data_attr_callback($matches) {
+		$content = preg_replace_callback('#data-[^\'"]+=[\'"]\Khttp://[^\'"]+#i', array($this, 'fixContent_src_callback'), $matches[0]);
 
 		return $content;
 	}